+++ bigbuy/__init__.py
@@ -13,7 +13,6 @@ from .exceptions import (
     BBError, BBResponseError, BBPackError, BBExportError, BBProductError, BBStockError,
     BBNoCarrierError, BBBankWireTooLowError, BBMoneyBoxTooLowError, BBTemporaryOrderError, BBOrderAlreadyExistsError,
     BBOrderTooLowError, BBIncorrectRefError, BBInvalidPaymentError, BBZipcodeFormatError, BBProductNotFoundError,
-    BBServerError, BBRateLimitError, BBValidationError,
-    wait_rate_limit,
+    BBServerError, BBRateLimitError, BBValidationError, BBWarehouseSplitError, BBShippingError,
 )
 from .rate_limit import RateLimit